[MODORDERS-1141]. Move code from PieceUpdateFlowUtil to PieceUpdateFlowInventoryManager
